Rename `sandbox::linux` module to `linux_namespace`
